move get_number_of_forms_of_all_types and test to analytics modules
