+++ corehq/ex-submodules/couchforms/tests/test_analytics.py
@@ -8,8 +8,9 @@ from couchforms.analytics import domain_has_submission_in_last_30_days, \
     app_has_been_submitted_to_in_last_30_days, update_analytics_indexes, \
     get_username_in_last_form_user_id_submitted, get_all_user_ids_submitted, \
     get_all_xmlns_app_id_pairs_submitted_to_in_domain, \
-    get_last_form_submission_for_user_for_app, get_number_of_submissions, get_form_analytics_metadata
-from couchforms.models import XFormInstance
+    get_last_form_submission_for_user_for_app, get_number_of_submissions, get_form_analytics_metadata, \
+    get_number_of_forms_of_all_types
+from couchforms.models import XFormInstance, XFormError
 
 
 class CouchformsAnalyticsTest(TestCase, DocTestMixin):
@@ -33,14 +34,16 @@ class CouchformsAnalyticsTest(TestCase, DocTestMixin):
                           app_id=cls.app_id, xmlns=cls.xmlns,
                           form={'meta': {'userID': cls.user_id, 'username': 'frank'}}),
         ]
-        for form in cls.forms:
+        cls.error_forms = [XFormError(domain=cls.domain)]
+        cls.all_forms = cls.forms + cls.error_forms
+        for form in cls.all_forms:
             form.save()
 
         update_analytics_indexes()
 
     @classmethod
     def tearDownClass(cls):
-        for form in cls.forms:
+        for form in cls.all_forms:
             form.delete()
 
     def test_domain_has_submission_in_last_30_days(self):
@@ -98,3 +101,9 @@ class CouchformsAnalyticsTest(TestCase, DocTestMixin):
         info = get_form_analytics_metadata(self.domain, self.app_id, self.xmlns)
         self.assertEqual(self.xmlns, info['xmlns'])
         self.assertEqual(2, info['submissions'])
+
+    def test_get_number_of_forms_of_all_types(self):
+        self.assertEqual(
+            get_number_of_forms_of_all_types(self.domain),
+            len(self.all_forms)
+        )
