+++ corehq/ex-submodules/couchforms/tests/test_dbaccessors.py
@@ -2,8 +2,7 @@ import datetime
 from django.test import TestCase
 
 from corehq.form_processor.tests.utils import FormProcessorTestUtils
-from couchforms.dbaccessors import get_forms_by_type, \
-    get_number_of_forms_by_type, get_number_of_forms_of_all_types, \
+from couchforms.dbaccessors import get_forms_by_type, get_number_of_forms_by_type, \
     get_form_ids_by_type, get_number_of_forms_all_domains_in_couch
 from couchforms.models import XFormInstance, XFormError
 
@@ -58,12 +57,6 @@ class TestDBAccessors(TestCase):
         self.assertEqual(
             get_number_of_forms_by_type(self.domain, 'XFormError'),
             len(self.xform_errors)
-        )
-
-    def test_get_number_of_forms_of_all_types(self):
-        self.assertEqual(
-            get_number_of_forms_of_all_types(self.domain),
-            len(self.xforms) + len(self.xform_errors)
         )
 
     def test_get_form_ids_by_type(self):
